Allow custom initial and observable states
We just return the state object as it is, if it is not an instance of
`Stete`.  However, there were comparisons to the state object (usually
the special case of `Stete.EQUILIBRIUM`) and there, the `==` caused an
error because it is overloaded for numpy array (needs `any` or `all`
method calls) so we also changes the `==` operators to `is`.